отрефакторил код
