+++ src/presenter/content-presenter.js
@@ -2,7 +2,7 @@ import ContentView from '../view/content-view.js';
 import MessageView from '../view/message-view.js';
 import WaypointPresenter from './waypoint-presenter.js';
 import SortContainerView from '../view/sort-container-view.js';
-import { FilterType, newWaypoint, SortType, UpdateType, UserAction } from '../const.js';
+import { FilterType, SortType, UpdateType, UserAction } from '../const.js';
 import SortModel from '../model/sort-model.js';
 import { remove, render } from '../framework/render.js';
 import TripInfoView from '../view/trip-info-view.js';
@@ -11,6 +11,7 @@ import NewPointPresenter from './new-point-presenter.js';
 import { filter } from '../utils/util-filter.js';
 import { nanoid } from 'nanoid';
 import LoadingView from '../view/loading-view.js';
+import { newWaypoint } from '../utils/util-waypoint.js';
 
 export default class ContentPresenter {
   #boardComponent = new ContentView();
@@ -49,15 +50,36 @@ export default class ContentPresenter {
     this.#filterModel.addObserver(this.#handleModelEvent);
   }
 
-  init() {
-    this.#renderContentContainer();
-  }
-
   get waypoints(){
     this.#filterType = this.#filterModel.filter;
     const waypoints = this.#waypointModel.humanizedWaypoints;
     const filteredWaypoints = filter[this.#filterType](waypoints);
     return [...this.#waypointModel.sortWaypoints(filteredWaypoints, this.#currentSortType)];
+  }
+
+  init() {
+    this.#renderContentContainer();
+  }
+
+  createWaypoint() {
+    const offersByType = newWaypoint.offersByType([...this.#waypointModel.offers]);
+
+    const destinations = this.#waypointModel.destinations;
+    this.#newWaypoint = {
+      id: nanoid(),
+      ...newWaypoint,
+      allDestinations: destinations,
+      allOffers: [...this.#waypointModel.offers],
+      destination: destinations[0],
+      offersByType: offersByType.offers,
+      dateFrom: new Date(),
+      dateTo: new Date(),
+      isFavorite: false,
+      allDestinationNames: this.#waypointModel.cities,
+    };
+    this.#currentSortType = SortType.DEFAULT;
+    this.#filterModel.setFilter(UpdateType.MAJOR, FilterType.EVERYTHING);
+    this.#newWaypointPresenter.init(this.#newWaypoint);
   }
 
   #renderSortings(){
@@ -67,16 +89,6 @@ export default class ContentPresenter {
       onSortTypeChange: this.#handleSortTypeChange});
     render(this.#sortingComponent, this.#sortingsContainer);
   }
-
-  #handleSortTypeChange = (sortType) => {
-    if(this.#currentSortType === sortType){
-      return;
-    }
-
-    this.#currentSortType = sortType;
-    this.#clearContentContainer();
-    this.#renderContentContainer();
-  };
 
   #renderPoints(waypoints){
     for (let i = 0; i < waypoints.length; i++) {
@@ -99,11 +111,6 @@ export default class ContentPresenter {
     render(this.#messageComponent, this.#contentContainer);
   }
 
-  #handleModeChange = () => {
-    this.#waypointPresentersList.forEach((presenter) => presenter.resetView());
-    this.#newWaypointPresenterList.forEach((presenter) => presenter.cancelAddPointClick());
-  };
-
   #renderTrip(){
     const tripModel = new TripModel(this.waypoints);
     let trip = tripModel.trip;
@@ -120,6 +127,67 @@ export default class ContentPresenter {
     this.#tripComponent = new TripInfoView({trip: trip});
     render(this.#tripComponent, this.#tripContainer,'AFTERBEGIN');
   }
+
+  #renderLoading(){
+    render(this.#loadingComponent, this.#contentContainer);
+  }
+
+  #clearContentContainer({resetSortType = false, resetFilterType = false} = {}) {
+    this.#waypointPresentersList.forEach((presenter) => presenter.destroy());
+    this.#waypointPresentersList.clear();
+    remove(this.#sortingComponent);
+    remove(this.#tripComponent);
+    remove(this.#loadingComponent);
+    this.#newWaypointPresenterList.clear();
+
+    if (resetSortType) {
+      this.#currentSortType = SortType.DAY;
+    }
+
+    if (resetFilterType) {
+      this.#filterType = FilterType.EVERYTHING;
+    }
+
+    if(this.#messageComponent){
+      remove(this.#messageComponent);
+    }
+  }
+
+  #renderContentContainer() {
+    if(this.#isLoading){
+      this.#renderLoading();
+      return;
+    }
+
+    this.#renderSortings();
+
+    render(this.#boardComponent, this.#contentContainer);
+
+    this.#renderPoints(this.waypoints);
+
+    this.#renderTrip();
+
+    const waypoints = this.waypoints;
+
+    if(waypoints.length === 0){
+      this.#renderMessage();
+    }
+  }
+
+  #handleSortTypeChange = (sortType) => {
+    if(this.#currentSortType === sortType){
+      return;
+    }
+
+    this.#currentSortType = sortType;
+    this.#clearContentContainer();
+    this.#renderContentContainer();
+  };
+
+  #handleModeChange = () => {
+    this.#waypointPresentersList.forEach((presenter) => presenter.resetView());
+    this.#newWaypointPresenterList.forEach((presenter) => presenter.destroy());
+  };
 
   #handleViewAction = (actionType, updateType, update) => {
     switch(actionType){
@@ -155,71 +223,4 @@ export default class ContentPresenter {
         break;
     }
   };
-
-  #clearContentContainer({resetSortType = false, resetFilterType = false} = {}) {
-    this.#waypointPresentersList.forEach((presenter) => presenter.destroy());
-    this.#waypointPresentersList.clear();
-    remove(this.#sortingComponent);
-    remove(this.#tripComponent);
-    remove(this.#loadingComponent);
-    this.#newWaypointPresenterList.clear();
-
-    if (resetSortType) {
-      this.#currentSortType = SortType.DAY;
-    }
-
-    if (resetFilterType) {
-      this.#filterType = FilterType.EVERYTHING;
-    }
-
-    if(this.#messageComponent){
-      remove(this.#messageComponent);
-    }
-  }
-
-  createWaypoint() {
-    const offersByType = newWaypoint.offersByType([...this.#waypointModel.offers]);
-
-    const destinations = this.#waypointModel.destinations;
-    this.#newWaypoint = {
-      id: nanoid(),
-      ...newWaypoint,
-      allDestinations: destinations,
-      allOffers: [...this.#waypointModel.offers],
-      destination: destinations[0],
-      offersByType: offersByType.offers,
-      dateFrom: new Date(),
-      dateTo: new Date(),
-      isFavorite: false,
-      allDestinationNames: this.#waypointModel.cities,
-    };
-    this.#currentSortType = SortType.DEFAULT;
-    this.#filterModel.setFilter(UpdateType.MAJOR, FilterType.EVERYTHING);
-    this.#newWaypointPresenter.init(this.#newWaypoint);
-  }
-
-  #renderContentContainer() {
-    if(this.#isLoading){
-      this.#renderLoading();
-      return;
-    }
-
-    this.#renderSortings();
-
-    render(this.#boardComponent, this.#contentContainer);
-
-    this.#renderPoints(this.waypoints);
-
-    this.#renderTrip();
-
-    const waypoints = this.waypoints;
-
-    if(waypoints.length === 0){
-      this.#renderMessage();
-    }
-  }
-
-  #renderLoading(){
-    render(this.#loadingComponent, this.#contentContainer);
-  }
 }
