+++ src/view/filter-container-view.js
@@ -2,7 +2,6 @@ import AbstractView from '../framework/view/abstract-view.js';
 
 const createFilterItemTemplate = (filter, currentFilterType) => {
   const {type, name, count} = filter;
-
   return (
     `<div class="trip-filters__filter">
           <input id="filter-${type}" 
@@ -46,5 +45,4 @@ export default class FilterContainerView extends AbstractView {
     evt.preventDefault();
     this.#handleFilterTypeChange(evt.target.value);
   };
-
 }
